+++ sqlite/package.json
@@ -5,19 +5,29 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "ts-node-dev --respawn --transpile-only src/server.ts"
+    "dev": "ts-node-dev -r dotenv/config --respawn --transpile-only src/server.ts"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "better-sqlite3": "^7.4.1",
+    "debug": "^4.3.1",
     "express": "^4.17.1",
+    "express-jwt": "^6.0.0",
+    "jsonwebtoken": "^8.5.1",
+    "nanoid": "^3.1.23",
+    "passport": "^0.4.1",
     "zod": "^3.2.0"
   },
   "devDependencies": {
     "@types/better-sqlite3": "^5.4.1",
     "@types/express": "^4.17.12",
-    "@types/node": "^15.12.2",
+    "@types/express-jwt": "^6.0.1",
+    "@types/jsonwebtoken": "^8.5.1",
+    "@types/nanoid": "^3.0.0",
+    "@types/passport": "^1.0.6",
+    "@types/passport-github2": "^1.2.4",
+    "dotenv": "^10.0.0",
     "ts-node": "^10.0.0",
     "ts-node-dev": "^1.1.6",
     "tsc": "^2.0.3",
